requirements.txt
pymodbus
pydantic
pyyaml
